feature/19 feat: 댓글 추가 Service 로직 구현
 - 댓글 추가 ServiceTest 추가
 - Location.calculateDistance 메소드 구현
 - Comment.distance의 자료형을 Long에서 Double로 변경
 - MemberNotFoundException 추가

+++ src/main/java/com/grasshouse/dorandoran/comment/domain/Comment.java
@@ -35,7 +35,7 @@ public class Comment {
 
     private String content;
 
-    private Long distance;
+    private Double distance;
 
     @OneToMany
     @JoinColumn(name = "COMMENT_LIKE_ID")
